refactor Term
